renamed eventprocessor to eventprocessorhost for consistency
