+++ eventprocessorhost/abstract_checkpoint_manager.py
