+++ eventprocessorhost/azure_blob_lease.py
@@ -4,7 +4,7 @@
 # -----------------------------------------------------------------------------------
 
 import json
-from eventhubsprocessor.lease import Lease
+from eventprocessorhost.lease import Lease
 class AzureBlobLease(Lease):
     """
     Azure Blob Lease
